# Optimizing common take operations