@nsmith- 2021-08-17
Is there a way I can see the windows build errors? It just says "failed" in the azure pipeline output..